* Moved the data processing up one level. * Add fixes in RoughAmplitude parameter updating.
